+++ packages/core/src/node/index.ts
@@ -7,6 +7,7 @@ import sharp from "sharp";
 import invariant, {
   Config,
   getCachePath,
+  getContentType,
   getImgParams,
   getImgSource,
   ImgParams,
@@ -70,6 +71,8 @@ export async function getImgResponse(request: Request, config: Config = {}) {
   }
   const params: ImgParams = paramsRes;
 
+  headers.set("Content-Type", getContentType(params.format));
+
   // Map src to location of the original image (fs or fetch)
   const sourceRes = config.getImgSource
     ? await config.getImgSource({ request, params })
@@ -128,10 +131,8 @@ export async function getImgResponse(request: Request, config: Config = {}) {
     const pipeline = sharp();
     if (params.format === "avif") {
       pipeline.avif();
-      headers.set("content-type", "image/avif");
     } else if (params.format === "webp") {
       pipeline.webp();
-      headers.set("content-type", "image/webp");
     }
 
     if (params.width && params.height) {